Moved winrm communication object to communication folder
